Move Apply and Clear back into FiltersBox
