+++ app/javascript/components/FiltersBox.js
@@ -1,27 +1,54 @@
 import React, { useState, useEffect } from 'react';
 import isEqual from 'lodash/isEqual';
+import Button from 'react-bootstrap/lib/Button';
 
 import Filters from './Filters';
-import {ApplyButton, ClearFilters, useApplyAndClear} from './ApplyAndClear';
+
+import { SearchContext } from './SearchPanel';
+
+/**
+ * Component that can be clicked to unselect filters
+ */
+export function ClearFilters(props) {
+  return (
+    <span
+      id={`clear-filters-${props.facetId}`}
+      className='clear-filters'
+      onClick={props.onClick}
+    >
+      CLEAR
+    </span>
+  );
+}
+
+/**
+ * Component for the "APPLY" button that can be clicked it to save selected
+ * filters for the current facet or facet accordion.
+ */
+export function ApplyButton(props) {
+  return (
+    <Button
+      id={props.id}
+      bsStyle='primary'
+      className={props.className}
+      onClick={props.onClick}
+    >
+    APPLY
+    </Button>
+  );
+}
 
 /**
  * Component for filter lists that have Apply and Clear
  */
 export default function FiltersBox(props) {
 
-  // State for reusable "APPLY" and "Clear" buttons.
-  // This uses a custom hook to encapsulate reusable state code and functions.
-  // The FacetsAccordionBox also uses this custom hook.
-  // It's like a Higher-Order Component, but for function components.
-  const {
-    canApply, setCanApply,
-    showClear,
-    appliedSelection,
-    selection,
-    updateSelections,
-    handleApplyClick,
-    clearFilters
-  } = useApplyAndClear();
+  const searchContext = useContext(SearchContext);
+
+  const [canApply, setCanApply] = useState(false);
+  const [showClear, setShowClear] = useState(false);
+  const [appliedSelection, setAppliedSelection] = useState([]);
+  const [selection, setSelection] = useState([]);
 
   useEffect(() => {
     setCanApply(!isEqual(selection, appliedSelection));
@@ -30,6 +57,50 @@ export default function FiltersBox(props) {
   useEffect(() => {
     setCanApply(false);
   }, [appliedSelection]);
+
+  /**
+   * Returns IDs of selected filters.
+   * Enables comparing current vs. applied filters to enable/disable APPLY button
+   */
+  function getCheckedFilterIds(facetBoxId) {
+    const checkedSelector = `#${facetBoxId} input:checked`;
+    const checkedFilterIds =
+      [...document.querySelectorAll(checkedSelector)].map((filter) => {
+        return filter.id;
+      });
+    return checkedFilterIds
+  }
+
+  function updateSelections(facetBoxId) {
+    const checkedFilterIds = getCheckedFilterIds(facetBoxId);
+    setSelection(checkedFilterIds);
+    setShowClear(checkedFilterIds.length > 0);
+  }
+
+  function handleApplyClick(event, facetId) {
+    const applyButtonClasses = Array.from(event.target.classList);
+
+    if (applyButtonClasses.includes('disabled')) return;
+
+    const checkedFilterIds = getCheckedFilterIds();
+
+    if (checkedFilterIds.length > 0) {
+      searchContext.facets[facetId] = checkedFilterIds.join(',');
+    } else {
+      delete searchContext.facets[facetId];
+    }
+
+    setAppliedSelection(checkedFilterIds);
+  }
+
+  function clearFilters(facetBoxId) {
+    const checkedSelector = `#${facetBoxId} input:checked`;
+    document.querySelectorAll(checkedSelector).forEach((checkedInput) => {
+      checkedInput.checked = false;
+    });
+
+    updateSelections();
+  }
 
   // TODO: Get opinions, perhaps move to a UI code style guide.
   //
